Introduce BaseNodeTest, start implementing the usage of cached bitcoinds in the nodeTest project
