+++ node-test/src/test/scala/org/bitcoins/node/UpdateBloomFilterTest.scala
@@ -5,12 +5,16 @@ import org.bitcoins.server.BitcoinSAppConfig
 import org.bitcoins.testkit.BitcoinSTestAppConfig
 import org.bitcoins.testkit.node.{
   NodeTestUtil,
-  NodeUnitTest,
+  NodeTestWithCachedBitcoindNewest,
   SpvNodeFundedWalletBitcoind
 }
-import org.scalatest.{BeforeAndAfter, FutureOutcome}
+import org.scalatest.{BeforeAndAfter, FutureOutcome, Outcome}
 
-class UpdateBloomFilterTest extends NodeUnitTest with BeforeAndAfter {
+import scala.concurrent.Future
+
+class UpdateBloomFilterTest
+    extends NodeTestWithCachedBitcoindNewest
+    with BeforeAndAfter {
 
   /** Wallet config with data directory set to user temp directory */
   implicit override protected def getFreshConfig: BitcoinSAppConfig =
@@ -19,7 +23,14 @@ class UpdateBloomFilterTest extends NodeUnitTest with BeforeAndAfter {
   override type FixtureParam = SpvNodeFundedWalletBitcoind
 
   def withFixture(test: OneArgAsyncTest): FutureOutcome = {
-    withSpvNodeFundedWalletBitcoind(test, None)
+    val outcome: Future[Outcome] = for {
+      bitcoind <- cachedBitcoindWithFundsF
+      outcome = withSpvNodeFundedWalletBitcoindCached(test,
+                                                      getBIP39PasswordOpt(),
+                                                      bitcoind)
+      f <- outcome.toFuture
+    } yield f
+    new FutureOutcome(outcome)
   }
 
   it must "update the bloom filter with a TX" in { param =>
